requirements.txt
fastapi
hypercorn
python-dotenv
tortoise
asyncpg
pydantic
argon2-cffi
pyjwt